+++ gulpfile.js
@@ -1,13 +1,13 @@
-var browserify = require('browserify');
 var gulp = require('gulp');
 var clean = require('gulp-clean');
 var sass = require('gulp-sass');
 var ts = require('gulp-typescript');
-var tsify = require('tsify');
 var source = require('vinyl-source-stream');
+var webpack = require('webpack-stream');
 
 var tsProject = ts.createProject('tsconfig.json', {
   typescript: require('typescript'),
+  target: "ES3",
 });
 
 var tslint = require('gulp-tslint');
@@ -28,7 +28,7 @@ gulp.task('transpile', function() {
 gulp.task('demo-copy', function() {
   gulp.src(['./demo/index.html', './demo/images/**/*'], { base: './demo' })
     .pipe(gulp.dest('docs'));
-})
+});
 
 // transpiles all the demo scss files
 gulp.task('demo-scss', function() {
@@ -39,17 +39,9 @@ gulp.task('demo-scss', function() {
 
 // generates all the demo files in the build directory
 gulp.task('demo', ['default', 'demo-copy', 'demo-scss'], function() {
-  return browserify({
-    basedir: '.',
-    debug: true,
-    entries: ['demo/index.tsx'],
-    cache: {},
-    packageCache: {},
-  })
-    .plugin(tsify, { typescript: require('typescript') })
-    .bundle()
-    .pipe(source('demo.js'))
-    .pipe(gulp.dest('docs'));
+  return gulp.src('demo/index.tsx')
+    .pipe(webpack( require('./webpack/webpack.demo.js') ))
+    .pipe(gulp.dest('docs/'));
 });
 
 // lints all the typescript source files