+++ test/main.tests.ts
@@ -1,0 +1,9 @@
+// the here maps tests make sure the scripts are loaded before proceeding
+import "./HEREMap.tests";
+
+// then we can complete the other tests
+import "./Circle.tests";
+import "./Marker.tests";
+import "./mixins/h-map-methods.tests";
+import "./utils/get-link.tests";
+import "./utils/mixin.tests";