+++ test/main.ts
@@ -1,4 +1,0 @@
-import "./HEREMap.tests";
-import "./Marker.tests";
-import "./mixins/h-map-methods.tests";
-import "./utils/get-link.tests";