+++ typings/packages/es6-map/es6-map.d.ts
@@ -1,4 +1,0 @@
-declare module "es6-map" {
-    const Map: MapConstructor;
-    export = Map;
-}