+++ typings/packages/reflect.ownkeys/reflect.ownkeys.d.ts
@@ -1,5 +1,0 @@
-declare module 'reflect.ownkeys' {
-    type PropertyKey = string | number | symbol;
-    function ownKeys(target: any): Array<PropertyKey>;
-    export = ownKeys;
-}